Remove workarounds, add HWND helper
